passado a norminette, so precisamos testar mapas invalidos
